Fix hash-syntax #'
From Marc N-W:

"The #' entry is also not quite right, semantically.  #' is just an
abbreviation for "syntax" as ' is an abbreviation for quote.  As
lexical syntax, it is not connected to syntax datums.  It is not
recognized by the syntax-case system, either.  It just happens that,
in the standard library, the identifier "syntax" is a keyword that,
when used, evaluates to a syntax object.

For example '#'x just evaluates to the list (syntax x)."